requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.116.1
grpcio==1.74.0
h11==0.16.0
pymongo==4.10.1
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-qdrant==0.2.0
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
portalocker==3.2.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
qdrant-client==1.15.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rq==2.4.1
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
zstandard==0.23.0
